fix: Styles in Profile
