+++ kit/UserProfile/ContentItems/SoundAndVideo/common/Icons/Arrow.tsx
