+++ kit/UserProfile/ContentItems/SoundAndVideo/common/Icons/Camera.tsx
