+++ kit/UserProfile/ContentItems/SoundAndVideo/common/Icons/MaxVolume.tsx
