+++ kit/UserProfile/ContentItems/SoundAndVideo/common/Icons/Microphone.tsx
