+++ kit/UserProfile/ContentItems/SoundAndVideo/common/Icons/MinVolume.tsx
