+++ kit/UserProfile/ContentItems/SoundAndVideo/common/VideoDevice.tsx
